Add exception handler and error response

+++ src/main/java/com/example/demo/product/services/DeleteProductService.java
@@ -1,5 +1,6 @@
 package com.example.demo.product.services;
 
+import com.example.demo.product.exceptions.ProductNotFoundException;
 import com.example.demo.product.repositories.ProductRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,8 +16,11 @@ public class DeleteProductService implements Command<Integer, Void> {
 
     @Override
     public ResponseEntity<Void> execute(Integer param) {
-        productRepository.deleteById(param);
-        return ResponseEntity.status(HttpStatus.OK).build();
+        if (productRepository.findById(param).isPresent()) {
+            productRepository.deleteById(param);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
+        throw new ProductNotFoundException();
     }
 
 }
